replaced "ranking" with "rating"
I think I'd been using the wrong word for the number representing the player's skill, such as 1500, or 1775.  I am now calling this "rating" instead of "ranking."  I think ranking means something more like you are rank 1 if you are on top of the leaderboard, rank 5 if you are 5th on the leaderboard.