potential fix for getting stuck at "game is starting"
also, displays a message if you get stuck, with diagnostic info, asking the user screenshot it and post in #panel-attack-bugs-features

it seems like the error was in processing match_start after the client recieved a Q message for Player 1's gpanel_buffer. it was resetting P1.gpanel_buffer to "", and now it doesn't.