workaround for music stopping
the problem seems to be we try to start playing the track again before it ends.  the play function won't do anything if the track is already playing.  Now, it will stop the track and play it, but this leads to looping not sounding quite right.  we slightly cut off the end of the track. (but i suppose this is better than having no music)