Fix network issues where game doesn't initialize when spectating
Together with shosoul we were able to determine that we were deleting game network data for the next game when cleaning
up the previous game.

We only want to delete all game messages prior to the "new game" network message.

Added a method to do that and added it to a couple places.
Also documented each network processing spot.
Tested with simulated network latency and was able to repro before and not afterwards.