Score and other Game Improvements
Scores now save and track
	# of garbage lines in VS Self
	Score in Endless
	Score in Time Attack

Scores also track the version and tries to handle old data and not overwrite new versions

Added a “Game” class to track the overall game state
Added a “Match” class to track the current match going on for example 1 time attack round, or 1 vs match
Moved all uses of the stack’s mode to the match’s mode

Analytics are now always captured on a per stack basis even if analytics are not enabled.
Analytics are only saved if analytics are enabled
The stack also records the overall states ONLY if the stack is local

Return to the last used mode when you exit a game
Allow exiting a local game if you pause and then press back or escape.

Testing Notes: Please test everything above plus general gameplay, spectating and playing net vs matches.