Patch: "You are all alone in the lobby" Bug
This patch fixes the bug where the game would say "You are all alone in the lobby :(".
This bug is very similar to the room init fail bug, but instead of leaving the select menu screen for the match start, you leave the select menu because the room closed before leaving the game over screen. This patch fixes this bug by adding the room closing message to the list of triggers to leave the game over screen.