Rollback Inputs implemented!
We now rollback inputs instead of garbage.
The benefit of this is there is virtually no "teleporting" of garbage in on the local stack
To accomplish this, we had to save everything in rollback that could be modified on the stack.
Also introduce a way to force rollbacks and check for divergence when debug mode is on.
Lots of other changes to support intense rollback.
The last major change is making garbage and new panel buffers generate locally on the client deterministically based on a seed. In the future this seed should change and come from the server for online games.