Table functions (#412)
Provides extensions to the table class in a new file table_util.
Migrated, deleted and/or renamed various functions in util.lua related to tables.
Added new file table_util_tests to guarantee the basic functionality of the methods (as most of them are still unused right now). I do intend to use all unused ones for the computerPlayer branch.