Tcp networking rework (#483)
Enable tcp-nodelay by default. This disables the Nagle algorithm which ensures player inputs are sent right away even if it consumes a little more bandwidth and overhead.

In our testing this has reduced the number of frames the player is behind from about 10-15 to 5-10. 
The input buffer is also smaller which means we should have less lag bursts.

Also added some extra server options that are off by default.