Telegraph: Standardize garbage attack time to PA (#488)
The timing of attacks in the telegraph build is slower than normal panel attack.

To fix this while making networking work well this change adjusts the timing of garbage landing.

First we make the telegraph owned by the player sending the attacks, this makes it so attacks don't "hang" over your side if your opponent is lagging.
If the opponent is laggy, the telegraph attacks can hover over your stack temporarily

Second, we send the garbage to the opponent 60 frames in the future like PA used to do. This avoids rollback almost all the time for both combo and chain attacks.

To adjust the timing so its the same as stable panel attack and not slow, we reduce the amount of time an attack animates and stays in the telegraph.

Thus combo attacks still take 150 frames like stable, and chains still take 60 frames after ending like stable.