Pop FX updates (#1091)
* Pop FX updates

This fixes a bunch of bugs with pop FX, and attack cards and also universally applies the same transparency algorithm we were using for cards to make popfx not get in the way of playing the game.

Fixed a typo in popfx_burstRotate
Fixed burst animation not circling card when "fadeburst" was used
Deleted bigTimer, randomMax, big_position  and popSize variables as they did nothing / unused
Updated the popfx radius calculation to closer match the inspirational games and be smoother
Updated the popfx to draw behind the cards and the attack animations behind the cards so you can always see the cards
Fixed a bug where the attack animation wouldn't come from the right spot if you did a combo chain on the same frame
Fixed a bug where the attack animation would use the original chain spot when you grew a chain instead of the new spot
Fixed a bug where the attack animation was offset by the frame offset
Fixed a bunch of logic to allow the attack animation to start at a different offset, but ended up keeping it the same due to the short time needed to get it to the telegraph
Updated consts order and max lag now better documents the source of the number
Updated character documentation

* Review feedback

Ironically the coordinates of telegraph where correct and adding 1 to combo looked the same even though chain was actually the one that went up 1. Fixed that and also documented the variables to be only attack draw position. Since push pushes the combo and chain seperately its not clean to do the work in telegraph, but seems fine to do in check matches for now since its also done for cards, at least we are explicit about the variable names now.

* rf