Reworked Shake Time Animation
Cleaned everything up to be way better and the values to be more consistent symmetrical and feel better.
Also made it when new shake increase over previous it is smoothed out a small amount.
With this I feel like no shake reduction can be the standard and a reduction of 2 can be good for those that need it.
I don't think people are really going to need the "classic" huge values as this is very similar to PPL which people seem to be okay with.